feat(mac-crafter): Reorganized commands in dedicated source code files for improved developer experience.
Had to rename Utils/Codesign.swift to Utils/Codesigning.swift to avoid duplicate module name compiler error.
